Move parseInstallLink to utils
